[ADD] Complete token delimiting recursion
Several features lacking for now : \n token is not sure to be recognized and & as a partial operator is not parsed properly yet. The function aren't entirely conform to the norm as they are subject to later modifications. The token delimiter gives back tokens as lstrs, they are made into t_tokens afterwards. It handles unclosed lines recursively, asking user for input and resuming current token delimitation until the corresponding closing quote is found, and so on until all quotes are closed. cstr input becomes the joined cstr of all user input for current command reading, so that the tokenization may continue and the proper user input gets added to personal and readline history.